Working component that can make any image pop up
I completed the image pop up component that can be applied to any image inside the app. Scale of the picture is fixed assuming all posters on the front page will the be same size. However, the images do not have other extra functionality such as movement on top of hovering, but this can be added.