About Us Page Developers Template
Added the "developers" section to the about us page, with 7 templates for the 7 developers we have. Each template contains an image for the person, as well as their name which will contain a link to their email, and their position in the project. The AboutUs.css file was also changed to a modular css file.