CSS Page Animations
I added simple animations for all the elements of the home and events page (about us page already had them) that happen when traveling to that page. The text fades in and components or images fade in and scale up slightly. I also added code to scroll to the top of every page when clicking on it.