Update CreateEventModal.js
Changed error message in CreateEventModal to mirror the error message in the SignUpModal. The text fields that the user fails to enter information in will be highlighted in red as well as have a red exclamation mark next to them. In addition, if the user clicks create event before filling out all of the text fields, an error message will pop up at the top, prompting the user to enter all information.